services/google-stt/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
google-cloud-speech==2.23.0
numpy==1.26.3
